move traits
