+++ toolkit/partner-chains-cli/src/permissioned_candidates.rs
@@ -1,8 +1,21 @@
+use ogmios_client::query_ledger_state::{QueryLedgerState, QueryUtxoByUtxoId};
+use ogmios_client::query_network::QueryNetwork;
+use ogmios_client::transactions::Transactions;
+use partner_chains_cardano_offchain::await_tx::FixedDelayRetries;
+use partner_chains_cardano_offchain::cardano_keys::CardanoPaymentSigningKey;
+use partner_chains_cardano_offchain::csl::NetworkTypeExt;
+use partner_chains_cardano_offchain::multisig::MultiSigSmartContractResult;
+use partner_chains_cardano_offchain::permissioned_candidates::{
+	get_permissioned_candidates, upsert_permissioned_candidates,
+};
 use serde::{Deserialize, Serialize};
+use sidechain_domain::{PermissionedCandidateData, UtxoId};
 use sp_core::crypto::AccountId32;
 use sp_core::{ecdsa, ed25519, sr25519};
 use sp_runtime::traits::IdentifyAccount;
 use std::fmt::{Display, Formatter};
+
+use crate::cmd_traits::{GetPermissionedCandidates, UpsertPermissionedCandidates};
 
 #[derive(Debug, Deserialize, Eq, PartialEq, PartialOrd, Ord, Serialize)]
 pub(crate) struct PermissionedCandidateKeys {
@@ -95,4 +108,35 @@ fn parse_sr25519(value: &str) -> Option<sr25519::Public> {
 fn parse_ed25519(value: &str) -> Option<ed25519::Public> {
 	let bytes = sp_core::bytes::from_hex(value).ok()?;
 	Some(ed25519::Public::from(<[u8; 32]>::try_from(bytes).ok()?))
+}
+
+impl<C: QueryLedgerState + QueryNetwork + Transactions + QueryUtxoByUtxoId>
+	UpsertPermissionedCandidates for C
+{
+	async fn upsert_permissioned_candidates(
+		&self,
+		await_tx: FixedDelayRetries,
+		genesis_utxo: UtxoId,
+		candidates: &[PermissionedCandidateData],
+		payment_signing_key: &CardanoPaymentSigningKey,
+	) -> anyhow::Result<Option<MultiSigSmartContractResult>> {
+		upsert_permissioned_candidates(
+			genesis_utxo,
+			candidates,
+			payment_signing_key,
+			self,
+			&await_tx,
+		)
+		.await
+	}
+}
+
+impl<C: QueryLedgerState + QueryNetwork> GetPermissionedCandidates for C {
+	async fn get_permissioned_candidates(
+		&self,
+		genesis_utxo: UtxoId,
+	) -> anyhow::Result<Option<Vec<PermissionedCandidateData>>> {
+		let network = self.shelley_genesis_configuration().await?.network.to_csl();
+		get_permissioned_candidates(genesis_utxo, network, self).await
+	}
 }
