+++ toolkit/partner-chains-cli/src/register/mod.rs
@@ -1,7 +1,15 @@
+use ogmios_client::query_ledger_state::{QueryLedgerState, QueryUtxoByUtxoId};
+use ogmios_client::query_network::QueryNetwork;
+use ogmios_client::transactions::Transactions;
+use partner_chains_cardano_offchain::await_tx::FixedDelayRetries;
+use partner_chains_cardano_offchain::cardano_keys::CardanoPaymentSigningKey;
+use partner_chains_cardano_offchain::register::run_register;
 use plutus_datum_derive::ToDatum;
 use secp256k1::PublicKey;
 use sidechain_domain::*;
 use std::{convert::Infallible, fmt::Display, str::FromStr};
+
+use crate::cmd_traits::Register;
 
 pub mod register1;
 pub mod register2;
@@ -56,5 +64,22 @@ pub struct StakePoolSigningKeyParam(pub ed25519_zebra::SigningKey);
 impl From<[u8; 32]> for StakePoolSigningKeyParam {
 	fn from(key: [u8; 32]) -> Self {
 		Self(ed25519_zebra::SigningKey::from(key))
+	}
+}
+
+impl<T> Register for T
+where
+	T: QueryLedgerState + Transactions + QueryNetwork + QueryUtxoByUtxoId,
+{
+	async fn register(
+		&self,
+		await_tx: FixedDelayRetries,
+		genesis_utxo: UtxoId,
+		candidate_registration: &CandidateRegistration,
+		payment_signing_key: &CardanoPaymentSigningKey,
+	) -> Result<Option<McTxHash>, String> {
+		run_register(genesis_utxo, candidate_registration, payment_signing_key, self, await_tx)
+			.await
+			.map_err(|e| e.to_string())
 	}
 }
